requirements.txt
APScheduler
blinker
certifi
charset-normalizer
click
Flask
gitdb
GitPython
idna
itsdangerous
Jinja2
MarkupSafe
psycopg2-binary
pytz
redis
requests
six
smmap
tzlocal
urllib3
Werkzeug
